created user jdbc dao and database settings property file

+++ src/main/java/com/epam/hotel/daos/Dao.java
@@ -2,10 +2,14 @@ package com.epam.hotel.daos;
 
 import java.util.List;
 
-public interface Dao <T> {
+public interface Dao<T> {
     T create(T entity);
-    boolean delete(int id);
+
+    boolean delete(long id);
+
     boolean update(T entity);
+
     List<T> getAll();
-    T getById(int id);
+
+    T getById(long id);
 }
